Fix route progress index clamp to an int.
int.clamp returns num on current Dart, so the maneuver cursor is converted before indexing the instruction list.

+++ lib/core/navigation/route_progress.dart
@@ -73,7 +73,7 @@ class RouteProgressCalculator {
     var instructionIndex = instructionHint.clamp(
       0,
       math.max(0, route.instructions.length - 1),
-    );
+    ).toInt();
     while (instructionIndex < route.instructions.length) {
       final instruction = route.instructions[instructionIndex];
       final toManeuver =
